requirements.txt
click==8.3.1
Cython==3.2.2
execnet==2.1.2
fix-protobuf-imports==0.1.7
iniconfig==2.3.0
packaging==25.0
pip-autoremove==0.10.0
pluggy==1.6.0
pprofile==2.2.0
protobuf==6.33.2
py==1.11.0
Pygments==2.19.2
pytest==9.0.2
pytest-forked==1.6.0
pyzmq==27.1.0
setuptools==80.9.0
typing_extensions==4.15.0
xxhash==3.6.0
